refactor : legacy 제거
