--- app/src/main/java/camp/nextstep/ManualHandlerMapping.java
@@ -1,39 +1,0 @@
-package camp.nextstep;
-
-import camp.nextstep.controller.*;
-import com.interface21.webmvc.servlet.mvc.tobe.HandlerMapping;
-import jakarta.servlet.http.HttpServletRequest;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
-import com.interface21.webmvc.servlet.mvc.asis.ForwardController;
-
-import java.util.HashMap;
-import java.util.Map;
-
-public class ManualHandlerMapping implements HandlerMapping {
-
-    private static final Logger log = LoggerFactory.getLogger(ManualHandlerMapping.class);
-
-    private final Map<String, Controller> controllers = new HashMap<>();
-
-    public void initialize() {
-//        controllers.put("/", new ForwardController("/index.jsp"));
-//        controllers.put("/login", new LoginController());
-//        controllers.put("/login/view", new LoginViewController());
-//        controllers.put("/logout", new LogoutController());
-//        controllers.put("/register/view", new RegisterViewController());
-//        controllers.put("/register", new RegisterController());
-
-        log.info("Initialized Handler Mapping!");
-        controllers.keySet()
-            .forEach(path -> log.info("Path : {}, Controller : {}", path, controllers.get(path).getClass()));
-    }
-
-    @Override
-    public Object getHandler(HttpServletRequest request) {
-        String requestURI = request.getRequestURI();
-        log.debug("Request Mapping Uri : {}", requestURI);
-        return controllers.get(requestURI);
-    }
-}
