+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/Value.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc.asis;
+package com.interface21.webmvc.servlet.mvc.tobe;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
