+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/config/HandlerScanConfig.java
@@ -1,8 +1,5 @@
 package com.interface21.webmvc.servlet.mvc.tobe.config;
 
-import com.interface21.webmvc.servlet.mvc.asis.Value;
-
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
